`Base`: `append!` convert length to `Int` before passing it on
Reduces the number of invalidations from 2061 to 2045 on running this
code:

```julia
struct I <: Integer end
Base.Int(::I) = 7
```

NB: `_growend!`, which the length gets passed on to, converts to `Int`
immediately anyway.

+++ base/array.jl
@@ -1348,7 +1348,7 @@ function append! end
 
 function append!(a::Vector{T}, items::Union{AbstractVector{<:T},Tuple}) where T
     items isa Tuple && (items = map(x -> convert(T, x), items))
-    n = length(items)
+    n = Int(length(items))::Int
     _growend!(a, n)
     copyto!(a, length(a)-n+1, items, firstindex(items), n)
     return a
